contrib: add vmtest-tool to test BPF programs
This patch adds the vmtest-tool subdirectory under contrib which tests
BPF programs under a live kernel using a QEMU VM.  It automatically
builds the specified kernel version with eBPF support enabled
and stores it under "~/.bpf-vmtest-tool", which is reused for future
invocations.

It can also compile BPF C source files or BPF bytecode objects and
test them against the kernel verifier for errors.  When a BPF program
is rejected by the kernel verifier, the verifier logs are displayed.

$ python3 main.py -k 6.15 --bpf-src assets/ebpf-programs/fail.c
BPF program failed to load
Verifier logs:
        btf_vmlinux is malformed
        0: R1=ctx() R10=fp0
        0: (81) r0 = *(s32 *)(r10 +4)
        invalid read from stack R10 off=4 size=4
        processed 1 insns (limit 1000000) max_states_per_insn 0 total_states 0 peak_states 0 mark_read 0

See the README for more examples.

The script uses vmtest (https://github.com/danobi/vmtest) to boot
the VM and run the program.  By default, it uses the host's root
("/") as the VM rootfs via the 9p filesystem, so only the kernel is
replaced during testing.

Tested with Python 3.9 and above.

contrib/ChangeLog:

	* bpf-vmtest-tool/.gitignore: New file.
	* bpf-vmtest-tool/README: New file.
	* bpf-vmtest-tool/bpf.py: New file.
	* bpf-vmtest-tool/config.py: New file.
	* bpf-vmtest-tool/kernel.py: New file.
	* bpf-vmtest-tool/main.py: New file.
	* bpf-vmtest-tool/pyproject.toml: New file.
	* bpf-vmtest-tool/tests/test_cli.py: New file.
	* bpf-vmtest-tool/utils.py: New file.
	* bpf-vmtest-tool/vm.py: New file.

Signed-off-by: Piyush Raj <piyushraj92739@gmail.com>